+++ CHANGELOG.md
@@ -1,1 +1,23 @@
 # Changelog
+
+## 0.4.0
+
+### Added
+
+* Added Memcache predefined backend
+* Added AioMemcache predefined backend
+
+## Misc
+
+* Update documentation
+* Update Ci checks
+
+## 0.3.0
+
+* AioRedis predefined backend
+
+## 0.2.0 
+
+* Base predefined CookieBackend
+* Redis predefined Backend
+